requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
django-countries==7.4.2
django-mysql==4.9.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
rstr==3.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
sqlparse==0.4.3
typing_extensions==4.4.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
